MTSDK-26 Send Typing Indicator. (#89)
* MTSDK-26 Add EventListener

- Add `indicateTyping` public api to send `typing` events from user.
- Ensure that `typing` indicator not sent more that once every 5 seconds.
- Extract commonality between EventMessage.kt and TextMessage.kt to BaseMessageProtocol.kt
- Add EventMessage.kt data class for `events` sent to Shyrka.
- Ensure that UI is notified about Outbound events only.
- Include eventType in StructuredMessageEvent.kt
- If duration is not present or null in agent TypingEvent, fallback to default value of 5000ms.
- Just to be on the safe side, use type Long for typing `duration` field instead of Int.
- Add KDocs for public API.

+++ transport/src/commonMain/kotlin/com/genesys/cloud/messenger/transport/core/events/Event.kt
@@ -9,5 +9,5 @@ sealed class Event {
      *
      *  @param durationInMilliseconds amount of time to display visual typing indicator in UI.
      */
-    data class Typing(val durationInMilliseconds: Int) : Event()
+    data class Typing(val durationInMilliseconds: Long) : Event()
 }
